Feature: Add material and mesh for rendering (#8)
* Fix imgui sync error

fixes #7

* Add mouseDragging

* Fix multiple rendering on main Window

fixes #7

* fix bug where the scenePanel still pan when dragging window

* Support mouse wheel input in viewports

* Support depth stencil for editor

Note: Runtime still has bug

* Simple event trigger runtime demo

* Fix naming typo

* Add 'TODOs' for renderTarget

* Add render target pool

* Restructure innoBase and innoInternal

* Add material and mesh

* Support Imgui Column weight to EditorGUILayout

* Add implicit cast for math types

+++ InnoInternal/Render/Bridge/VeldridCommandList.cs
@@ -1,4 +1,6 @@
 using InnoBase;
+using InnoBase.Graphics;
+using InnoBase.Math;
 using InnoInternal.Render.Impl;
 
 using Veldrid;
@@ -55,7 +57,7 @@ internal class VeldridCommandList : ICommandList
     public void SetIndexBuffer(IIndexBuffer indexBuffer)
     {
         if (indexBuffer is VeldridIndexBuffer veldridIB)
-            inner.SetIndexBuffer(veldridIB.inner, IndexFormat.UInt16); // or UInt32 based on your index data
+            inner.SetIndexBuffer(veldridIB.inner, IndexFormat.UInt32);
     }
 
     public void SetResourceSet(int setIndex, IResourceSet resourceSet)
